Refactor imports to move search from libs.data to libs.functions for better organization
