--- libs/data/__init__.py
@@ -1,5 +1,0 @@
-"""
-データ操作モジュール
-
-- `libs.data.search`: DB検索
-"""
