+++ libs/functions/__init__.py
@@ -2,6 +2,7 @@
 共通処理モジュール
 
 - `libs.functions.lookup`: 情報取得
+- `libs.functions.search`: DB検索
 - `libs.functions.message`: メッセージ
 - `libs.functions.compose`: メッセージ生成
 - `libs.functions.tools`: 外部ツール用モジュール
