+++ libs/utils/dictutil.py
@@ -6,10 +6,9 @@ import logging
 from typing import TYPE_CHECKING, Any, Protocol
 
 import libs.global_value as g
-from libs.data import search
 from libs.domain.command import CommandParser
 from libs.domain.placeholder import PlaceholderBuilder
-from libs.functions import lookup
+from libs.functions import lookup, search
 from libs.types import ChannelType
 from libs.utils import formatter
 from libs.utils.timekit import ExtendedDatetime as ExtDt
